# openapi.yaml
openapi: 3.0.0
info:
  title: TrustLens API
  version: 1.0.0
  description: AI-powered authenticity verification platform for e-commerce
  contact:
    name: TrustLens Support
    email: support@trustlens.com
    url: https://trustlens.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.trustlens.com
    description: Production server
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /api/v1/uploads:
    post:
      summary: Upload file for analysis
      description: Upload an image or video file for authenticity verification
      tags:
        - Uploads
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Image or video file to analyze
                merchant_id:
                  type: string
                  description: Optional merchant identifier
                metadata:
                  type: object
                  description: Additional metadata about the upload
              required:
                - file
      responses:
        '200':
          description: File uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400':
          description: Invalid file or request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: File too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/auth/login:
    post:
      summary: User login
      description: Authenticate user and return access token
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  example: "password123"
              required:
                - email
                - password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for merchant/enterprise access
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authenticated users
  schemas:
    Error:
      type: object
      required:
        - error
        - message
        - code
      properties:
        error:
          type: string
          description: Error type
          example: "Validation Error"
        message:
          type: string
          description: Human-readable error message
          example: "The uploaded file is not a valid image format"
        code:
          type: string
          description: Machine-readable error code
          example: "INVALID_FILE_FORMAT"
        details:
          type: object
          description: Additional error details
        timestamp:
          type: string
          format: date-time
        path:
          type: string
          example: "/api/v1/uploads"
        method:
          type: string
          example: "POST"
    UploadResponse:
      type: object
      required:
        - upload_id
        - status
        - analysis_url
      properties:
        upload_id:
          type: string
          description: Unique identifier for the upload
          example: "clh5x8k2p0000qh8v4g8v4g8v"
        status:
          type: string
          enum: [pending, processing, completed, failed]
          description: Current status of the upload
          example: "pending"
        analysis_url:
          type: string
          format: uri
          description: URL to retrieve analysis results
          example: "https://api.trustlens.com/v1/analysis/clh5x8k2p0000qh8v4g8v4g8v"
        estimated_completion:
          type: string
          format: date-time
          description: Estimated completion time
          example: "2024-08-12T12:25:00Z"
...
    AuthResponse:
      type: object
      required:
        - access_token
        - token_type
        - expires_in
      properties:
        access_token:
          type: string
          description: JWT access token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          description: Token type
          example: "Bearer"
        expires_in:
          type: integer
          description: Token expiration time in seconds
          example: 3600
        refresh_token:
          type: string
          description: Refresh token for obtaining new access tokens
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        user:
          type: object
          properties:
            id:
              type: string
              example: "clh5x8k2p0000qh8v4g8v4g8v"
            email:
              type: string
              format: email
              example: "user@example.com"
            role:
              type: string
              enum: [admin, merchant, reviewer, support]
              example: "merchant"
...